requirements.txt
playwright==1.44.0
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2
python-telegram-bot==21.3
